Compare based on decompilation not raw binaries.
Builds of the same source code on different machines can result in
unpredictable differences in the compiled output. It is probably not
impossible to get this tool working with the original binary
comparison approach, but I am out of time.

Instead of Mono.Cecil, the tool now uses ILDASM to decompile each binary,
and then uses regular expression replacements to canonicalise some build-specific
tokens before computing the MD5 hash of the decompiled output. As a result
the tool is now A LOT slower. However, it now works. :)