Anonymizations: Extra .data cil; Referenced assembly version number. Speed optimization
Anonymization: Extra .data cil - sometimes an extra line of the form "" was added to the il file.

Anonymization: Referenced assembly version number - it may not be obvious why one would like to anonymize this, but an upgrade in AssemblyInfo of a referenced library causes the referencing DLL to yield a change in MD5 hash of the decompiled il file. Did not like this.

Speed optimization by means of anonymizing the il file line by line, which enabled a preliminary check whether or not a line qualifies for anonymization. This resulted in speed improvement from 260s to 6s for a 8MB dll file.